Fixed currently elapsing player cooldowns not updating
